requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0

# AI Integration
google-cloud-aiplatform==1.38.1
vertexai==1.38.1

# Environment Configuration
python-dotenv==1.0.0